backend/pyproject.toml
[tool.poetry]
name = "talktodata-backend"
version = "0.1.0"
description = "Talk to Data - AI-powered Text-to-SQL Analytics Platform"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
langchain = "^0.1.4"
langchain-anthropic = "^0.1.4"
langchain-community = "^0.0.16"
google-cloud-bigquery = "^3.14.1"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
celery = "^5.3.6"
loguru = "^0.7.2"
sentry-sdk = {extras = ["fastapi"], version = "^1.39.2"}
prometheus-client = "^0.19.0"
httpx = "^0.26.0"
python-dotenv = "^1.0.0"
cryptography = "^42.0.0"
sqlparse = "^0.4.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
black = "^23.12.1"
ruff = "^0.1.11"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
